requirements.txt
google-api-python-client==1.6.4
google-auth==1.3.0
google-auth-httplib2==0.0.3
speedtest-cli==1.0.7
